Project Overview
Brief overview of the various aspects of the project.
NOTE: This file is still not finished.  Had trouble getting the emulator to work in Android Studio and couldn't view app design in the design editor.  Once the emulator is working and screenshots of the different menus within the app can be taken, the "Android App" section of this file will be updated and will display the various menus of the app as well as explain the purpose and functionality of each menu.